Refactor FamilyScrollView and FamilyWrapperView
Refactor the underlying implementation of FamilyScrollView & FamilyWrapperView to work better with collection views transitioning to new layouts and overall animation operability.

Also fixes a minor issue with the content size being too big for the FamilyScrollView if the contents are lesser in height than the window that the FamilyScrollView belongs to.