Fix scrolling bug with single view controller
Fixes a bug that the underlying scroll view couldn't be scrolled if there was only one view controller.

Now it will opt-out of doing proxy scrolling if there is only one view controller added. This is done because of performance. There is no need to invoke the layout algorithm if there is only one scroll view in the list of views.