Refactor the layout algorithm on macOS (#52)
Refactor the layout algorithm on macOS to increase performance by using cached results instead of computed values.

```
This PR refactors the layout algorithm on macOS to increase performance. It does this by introducing an additional more lightweight algorithm that uses the cached results from the old algorithm.

So it works like this; if there are no cached results, it will invoke the old algorithm and create an internal cache from the results of that algorithm. The next layout pass will then use the results of the cache in order to layout the views in linear order. Cache invalidation happens if any arguments tied to the algorithm is changed, like changing the spacing between views. The cache is also invalidated when views are added or remove from the view hierarchy. Last but not least, if a view changes its high, the cache is also invalidated and then re-cached.
```