Add contentOffset observer
This observer will correct the current content offset of horizontal scroll views. The content offset should always be 0 for horizontal scroll views but when scrolling a collection of horizontal scroll views that recently got a frame, the system can set the content offset to a different value as the frame and content size are currently out-of-sync. On tvOS, the system wants to perform a scroll in order to make the current content visible, this is now prevented by using the observer.

This can also be fixed by manually adding the following to the scroll views delegate.

```
func scrollViewWillEndDragging(_ scrollView: UIScrollView, withVelocity velocity: CGPoint, targetContentOffset: UnsafeMutablePointer<CGPoint>) {
  targetContentOffset.pointee.y = 0
}
```

However, this needs to be on the client and not inside of the framework which makes it error-prone as the developer implementing Family needs to know about this specific issue and more importantly, needs to know to implement it every time there is a need to add a horizontal scroll view in the current UI composition. This is a best effort solution to work around the system correcting the content offset even if the frame and content size are set to a sane and valid value.

The issue was discovered on tvOS.