Add support for sprint animation updates
By adding a `CASpringAnimation` to the view that updates, it is now possible to
animate the entire view hierarchy using a spring animation.
Determining which animation should be used is now more accurate as the animation
will be resolved inside the observer. This makes it more customizable as individual
components can use different types of animations when their layout attributes change.